+++ WebMarket/WebMarket.Data.Model/Product.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WebMarket.Data.Model
+{
+    public class Product
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public decimal Price { get; set; }
+        public string Descript { get; set; }
+
+        public override string ToString()
+        {
+            return $"Id={Id}, Name={Name}, Price={Price}, Des={Descript}";
+        }
+    }
+}
