+++ WebMarket/WebMarket.Data.Model/ShoppingItem.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WebMarket.Data.Model
+{
+    public class ShoppingItem
+    {
+        public int ProductId { get; set; }
+        public int Count { get; set; }
+        public decimal Price { get; set; }
+
+        public decimal TotalPrice
+        {
+            get
+            {
+                return this.Count * this.Price;
+            }
+        }
+    }
+}
