+++ WebMarket/WebMarket.DebugConsole/App.config
